refactor close editors on startup and use only at startup
