+++ src/extension-utilities.ts
@@ -1,4 +1,4 @@
-import { commands, Position, Range, TextDocument, TextEditor, TextEditorEdit, window } from 'vscode';
+import { commands, env, Position, Range, TextDocument, TextEditor, TextEditorEdit, window } from 'vscode';
 
 export class ExtensionUtilities {
   /**
@@ -69,6 +69,38 @@ export class ExtensionUtilities {
       }
     } catch (err) {
       console.error(err);
+    }
+  }
+
+  static async closeEditorsOnStartup(cacheBasePath: string): Promise<void> {
+    /**
+     * This nonsense is necessary to close all open OOXML Viewer files on startup, because of this issue with VS Code:
+     * https://github.com/microsoft/vscode/issues/15178
+     * on startup, VS Code only recognizes the open tab in workspace.textDocuments, giving it a length of 1, so
+     * ExtensionUtilities.closeEditors(workspace.textDocuments) only closes one text editor, so we loop through all the text
+     * editors to make them active and close each active editor.
+     *
+     * And we use the clipboard to get the file name, because window.activeTextEditor.document is undefined with
+     * unsupported or binary files, so we can't use window.activeTextEditor.document.fileName
+     * https://github.com/Microsoft/vscode/issues/2582#issuecomment-246692860
+     *
+     */
+    let fileName: string | undefined;
+    const fileNames: string[] = [];
+    await commands.executeCommand('workbench.action.files.copyPathOfActiveFile');
+    fileName = await env.clipboard.readText();
+
+    while (!fileNames.includes(fileName)) {
+      fileNames.push(fileName);
+
+      await commands.executeCommand('workbench.action.nextEditor');
+
+      await commands.executeCommand('workbench.action.files.copyPathOfActiveFile');
+      fileName = await env.clipboard.readText();
+
+      if (fileName.toLowerCase().startsWith(cacheBasePath.toLowerCase())) {
+        await commands.executeCommand('workbench.action.closeActiveEditor');
+      }
     }
   }
 }
